backend/requirements.txt
Django==4.2.13
djangorestframework-simplejwt
django_allauth
dj_rest_auth
requests
python-dotenv
django-cors-headers
boto3
gunicorn
Pillow
channels
